Add a "No wallpaper" wallpaper option.
The simplest implementation is with a black 1x1px image,
but it could be implemented in a more efficient way in the future.
(i.e. the black image is an implementation detail of the
"no wallpaper" goal).

On Nexus One (OLED) this "no wallpaper" produces
45mA current saving compared to the "CM Earth" image wallpaper
when display is set to max brightness, and 20mA saving on low brightness.

Change-Id: Id712314d4e0808145f2180001bfc4dbeea766a46