Add option to set lock screen wallpaper on N+.
When clicking "Set wallpaper" in the cropper, a dialog pops up asking
which wallpaper you want to set. Though the implemenation does not yet
do anything, the dialog correctly shows up in the following cases:
- Opening a photo in Photos, and clicking "Use as" > "Photos"
- Settings > Display > Wallpaper, as long as you use our cropper
- Long-pressing the home screen > "Wallpapers" > "My photos"
- Long-pressing the home screen > built in wallpapers

The most notable case where it doesn't work yet is for live wallpapers,
because that goes through a different activity than the cropper.
Change-Id: Ia51d9e49fe5cc5845c006bafe3c300095652e7c6