Wallpaper dialog is cancelable.
If you're setting a wallpaper and are prompted to choose between lock
screen and home screen, you can now cancel the dialog. This was
previously disallowed because the "Set wallpaper" action bar and the
wallpaper tiles are hidden when the dialog is shown; now they are
reshown if the dialog is canceled.

Bug: 27484728
Change-Id: Ibe98c1320ebe185a4cb16254b5dec5105db4f1b7